Fix -Wsign-conversion in _nx_crypto_rsa_pss_sign salt generation
NX_CRYPTO_RAND() returns a signed int; masking it with 0xFFu implicitly
converts it to unsigned, which upstream CI rejects (-Werror -Wconversion).
Cast explicitly, matching the (UINT)NX_CRYPTO_RAND() idiom used in
nx_crypto_drbg.c and nx_crypto_huge_number.c.

Same fix as pushed to upstream PR #399 (commit e973e863).

+++ crypto_libraries/src/nx_crypto_rsa.c
@@ -873,7 +873,7 @@ static const UCHAR _pss_zero8[8] = {0, 0, 0, 0, 0, 0, 0, 0};
      * as the existing signature paths on this device. */
     for (i = 0u; i < s_len; i++)
     {
-        salt[i] = (UCHAR)(NX_CRYPTO_RAND() & 0xFFu);
+        salt[i] = (UCHAR)((UINT)NX_CRYPTO_RAND() & 0xFFu);
     }
 
     /* Steps 5-6 – H = Hash(0x00^8 || mHash || salt). Write directly into
